Extract sandbox scoring for custom checks
Move the helpers into a neutral module so custom checks can reuse them without depending on EvalConfig. Keep existing re-exports and scoring semantics.
