+++ modal_training_gym/__init__.py
@@ -14,7 +14,7 @@ _EXPORTS = {
     "AudioEvalRowResult": ("modal_training_gym.common.eval", "AudioEvalRowResult"),
     "ImageEvalRowResult": ("modal_training_gym.common.eval", "ImageEvalRowResult"),
     "Sample": ("modal_training_gym.common.sample", "Sample"),
-    "extract_code": ("modal_training_gym.common.eval", "extract_code"),
+    "extract_code": ("modal_training_gym.common.sandbox_scoring", "extract_code"),
     "HarborEval": ("modal_training_gym.common.eval", "HarborEval"),
     "ensure_endpoint": ("modal_training_gym.common.endpoint", "ensure_endpoint"),
     "endpoint_chat": ("modal_training_gym.common.endpoint", "endpoint_chat"),
@@ -91,7 +91,10 @@ _EXPORTS = {
         "modal_training_gym.train_recipes.slime_recipe",
         "Qwen3_ASR_1_7b_Recipe",
     ),
-    "score_in_sandbox": ("modal_training_gym.common.eval", "score_in_sandbox"),
+    "score_in_sandbox": (
+        "modal_training_gym.common.sandbox_scoring",
+        "score_in_sandbox",
+    ),
     "SlimeRecipe": ("modal_training_gym.train_recipes.slime_recipe", "SlimeRecipe"),
     "ToolCall": ("modal_training_gym.common.models", "ToolCall"),
     "TrainConfig": ("modal_training_gym.common.train", "TrainConfig"),
